feat(results): populate Participant Voice with 22 curated quotes; add CRP-2026 variant
## What changed since the initial PR commit

1. **22 curated quotes published** (3 praise + 11 insight + 1 critique +
   7 essay) selected from the 47 candidates surfaced by the local
   curation workflow. One essay (Pc4009c) was lightly edited to remove
   a false-positive employer-self-id pattern flagged by the PII scan.

2. **Refactored to shared component pattern** (matches SampleContent /
   DescriptiveContent / etc): `src/components/results/participant-voice/
   ParticipantVoiceContent.tsx` owns the layout; the route shells just
   pick a variant ('live' or 'crp') and pass the data.

3. **CRP-2026 variant added**: `/results/crp-2026/participant-voice/`
   exists alongside `/results/participant-voice/`. Both currently read
   the same JSON because the data source IS the frozen CRP-2026 N=200
   public dataset; the variant config differs only in title/intro/links
   so each fits its own results-overview neighborhood.

4. **Nav + sitemap** wired for both routes.

## Curation outcomes (with rationale in local claude_note column)

ESSAYS (5 verbatim + 1 edit):
  - P027331  AI/ROI ambivalence in software firm
  - P044416  GOLD: generational/Boomer leadership pattern
  - Pb17105  meta-critique of adoption-maturity assessments
  - P6069a2  sector transition (healthcare → education) gap
  - Pc4009c  EDIT: 'fintech SaaS' lowercased, employer-self-id phrasing
             rewritten while preserving substance
  - P0087c8  middle-management bottleneck pattern
  - P467723  construction industry / 'February morning in Ohio'

INSIGHTS (11 verbatim) including:
  - Pf7b010  75-year-old manufacturer, AI fear
  - Pb622e6  'still running green screens for our core sales system'
  - Pf8d6f6  'budget has increased 10-fold'
  - P7d2044  '25% of our workforce... still gettigg by it done. Great survey.'
  - Pbcf22a  'even the most prepared companies are still not quite ready'

CRITIQUE (1):
  - P7b3146  distributed-vs-centralized decision-making as missing dimension

PRAISE (3):
  - Pd672eb  'Great questionset!'
  - P40721f  'No, but thank you! :)'  (with emoji - shows affective range)
  - Pd873e9  'It's all covered, all clear'

## Privacy reaffirmed

All 22 published quotes passed PII scan (0 high-risk). Only Pc4009c
was flagged (employer-self-id pattern) and was reviewed manually -
the flagged 'Fintech SaaS' was a sector descriptor not a company name;
the EDIT version removes the regex trigger while keeping the substance.

Pseudonymous IDs (P + 6 chars of Qualtrics response hash) replace any
direct identifier. Anyone wanting to verify can match the IDs against
their own extraction from the public CSV.

## Live vs CRP

Today they show identical content because the only data is the frozen
CRP-2026 public dataset. Future iterations of the survey will let the
'live' variant diverge as new respondents are folded in; the shared
component already supports that without further refactoring.
